% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationCurves-package.R
\docType{package}
\name{CalibrationCurves-package}
\alias{CalibrationCurves}
\alias{CalibrationCurves-package}
\title{CalibrationCurves: Calibration Performance}
\description{
Plots calibration curves and computes statistics for assessing calibration performance. See Lasai et al. (2025) \doi{10.48550/arXiv.2503.08389}, De Cock Campo (2023) \doi{10.48550/arXiv.2309.08559} and Van Calster et al. (2016) \doi{10.1016/j.jclinepi.2015.12.005}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bavodc.github.io/websiteCalibrationCurves/}
}

}
\author{
\strong{Maintainer}: Bavo De Cock \email{bavo.decock@kuleuven.be}

Authors:
\itemize{
  \item Lasai Barrenada
  \item Daan Nieboer
  \item Ben Van Calster
  \item Ewout Steyerberg
  \item Yvonne Vergouwe
}

}
\keyword{internal}
