% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{ci_pi_labels}
\alias{ci_pi_labels}
\title{Generate CI/PI labels based on confidence level}
\usage{
ci_pi_labels(cl.level)
}
\arguments{
\item{cl.level}{The confidence level (between 0 and 1)}
}
\value{
Named vector with "ci" and "pi" labels formatted as percentages
}
\description{
Generate CI/PI labels based on confidence level
}
\keyword{internal}
