% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{\%<=\%}
\alias{\%<=\%}
\title{Infix operator to run background jobs}
\usage{
lhs \%<=\% rhs
}
\arguments{
\item{lhs}{the object that the rhs value is assigned to}

\item{rhs}{the value you want to assign to lhs}
}
\value{
prints the ID of the background job in the console and, once completed, the value of lhs is assigned to rhs
}
\description{
This infix operator can be used to create a background job in RStudio/Posit and, once completed, the value of rhs is
assigned to lhs.
}
\examples{
# Can only be executed in Rstudio
\dontrun{x \%<=\% rnorm(1e7)}
}
