% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{LibraryM}
\alias{LibraryM}
\title{Function to load multiple packages at once}
\usage{
LibraryM(...)
}
\arguments{
\item{...}{the packages that you want to load}
}
\value{
invisible \code{NULL}
}
\description{
Function to load multiple packages at once
}
\examples{
LibraryM(CalibrationCurves)
}
