% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\name{print.CalibrationCurve}
\alias{print.CalibrationCurve}
\title{Print function for a CalibrationCurve object}
\usage{
\method{print}{CalibrationCurve}(x, ...)
}
\arguments{
\item{x}{an object of type CalibrationCurve, resulting from \code{\link{val.prob.ci.2}}.}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original \code{CalibrationCurve} object is returned.
}
\description{
Prints the call, confidence level and values for the performance measures.
}
\seealso{
\code{\link{val.prob.ci.2}}
}
