% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\name{print.ClusteredCalibrationCurve}
\alias{print.ClusteredCalibrationCurve}
\title{Print function for a ClusteredCalibrationCurve object}
\usage{
\method{print}{ClusteredCalibrationCurve}(x, ...)
}
\arguments{
\item{x}{an object of type ggplotCalibrationCurve, resulting from \code{\link{valProbggplot}}.}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original \code{ggplotCalibrationCurve} object is returned.
}
\description{
Prints the ggplot, call, confidence level and values for the performance measures.
}
\seealso{
\code{\link{valProbggplot}}
}
