% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\name{print.CompRisksCalibrationCurve}
\alias{print.CompRisksCalibrationCurve}
\title{Print function for a CompRisksCalibrationCurve object}
\usage{
\method{print}{CompRisksCalibrationCurve}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CompRisksCalibrationCurve}, resulting
from \code{\link{valProbCompRisks}}.}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original object, invisibly.
}
\description{
Print function for a CompRisksCalibrationCurve object
}
\seealso{
\code{\link{valProbCompRisks}}
}
