% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\name{print.GeneralizedCalibrationCurve}
\alias{print.GeneralizedCalibrationCurve}
\title{Print function for a GeneralizedCalibrationCurve object}
\usage{
\method{print}{GeneralizedCalibrationCurve}(x, ...)
}
\arguments{
\item{x}{an object of type GeneralizedCalibrationCurve, resulting from \code{\link{genCalCurve}}.}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original \code{GeneralizedCalibrationCurve} object is returned.
}
\description{
Prints the call, confidence level and values for the performance measures.
}
\seealso{
\code{\link{genCalCurve}}
}
