% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\name{print.MulticlassCalibrationCurve}
\alias{print.MulticlassCalibrationCurve}
\title{Print function for a MulticlassCalibrationCurve object}
\usage{
\method{print}{MulticlassCalibrationCurve}(x, ...)
}
\arguments{
\item{x}{an object of class \code{MulticlassCalibrationCurve}, resulting
from \code{\link{valProbMulticat}}.}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original object, invisibly.
}
\description{
Print function for a MulticlassCalibrationCurve object
}
\seealso{
\code{\link{valProbMulticat}}
}
