% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\name{print.RecalibratedPredictions}
\alias{print.RecalibratedPredictions}
\title{Print function for a RecalibratedPredictions object}
\usage{
\method{print}{RecalibratedPredictions}(x, ...)
}
\arguments{
\item{x}{an object of class \code{RecalibratedPredictions}, resulting
from \code{\link{recalibrate}}.}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original object, invisibly.
}
\description{
Print function for a RecalibratedPredictions object
}
\seealso{
\code{\link{recalibrate}}
}
