% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printFunction.R
\name{print.SurvivalCalibrationCurve}
\alias{print.SurvivalCalibrationCurve}
\title{Print function for a SurvivalCalibrationCurve object}
\usage{
\method{print}{SurvivalCalibrationCurve}(x, ...)
}
\arguments{
\item{x}{an object of type SurvivalCalibrationCurve, resulting from \code{\link{valProbSurvival}}.}

\item{...}{arguments passed to \code{\link{print}}}
}
\value{
The original \code{SurvivalCalibrationCurve} object is returned.
}
\description{
Print function for a SurvivalCalibrationCurve object
}
\seealso{
\code{\link{valProbSurvival}}
}
