% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val.prob.ci.2.R
\name{val.prob.ci.2}
\alias{val.prob.ci.2}
\title{Calibration performance}
\usage{
val.prob.ci.2(p, y, logit, group, weights = rep(1, length(y)), normwt = F,
  pl = T, smooth = c("loess", "rcs", F), CL.smooth = "fill", CL.BT = F,
  nr.knots = 5, logistic.cal = F, xlab = "Predicted probability",
  ylab = "Observed proportion", xlim = c(-0.02, 1), ylim = c(-0.15, 1), m,
  g, cuts, emax.lim = c(0, 1), legendloc = c(0.5, 0.27), statloc = c(0,
  0.85), dostats = T, roundstats = 2, riskdist = "predicted",
  cex = 0.75, cex.leg = 0.75, connect.group = F, connect.smooth = T,
  g.group = 4, evaluate = 100, nmin = 0, d0lab = "0", d1lab = "1",
  cex.d01 = 0.7, dist.label = 0.04, line.bins = -0.05,
  dist.label2 = 0.03, cutoff, las = 1, length.seg = 1, y.intersp = 1,
  col.ideal = "grey", lwd.ideal = 1, ...)
}
\arguments{
\item{p}{
predicted probability
}

\item{y}{
vector of binary outcomes
}

\item{logit}{
predicted log odds of outcome.  Specify either \code{p} or \code{logit}.
}

\item{group}{
a grouping variable.  If numeric this variable is grouped into
\code{g.group} quantile groups (default is quartiles).  Set \code{group=TRUE} to
use the \code{group} algorithm but with a single stratum for \code{val.prob}.
}

\item{weights}{
an optional numeric vector of per-observation weights (usually frequencies),
used only if \code{group} is given.
}

\item{normwt}{
set to \code{TRUE} to make \code{weights} sum to the number of non-missing observations.
}

\item{pl}{\code{TRUE} to plot the calibration curve(s). If \code{FALSE} no calibration curves will be plotted,
but statistics will still be computed and outputted.}

\item{smooth}{\code{"loess"} generates a flexible calibration curve based on \code{\link{loess}},
\code{"rcs"} generates a calibration curves based on restricted cubic splines (see \code{\link{rcs}} and
\code{\link[Hmisc]{rcspline.plot}}), \code{FALSE} suppresses the flexible curve. We recommend to use loess unless N is large,
 for example N>5000. Default is \code{"loess"}.}

\item{CL.smooth}{\code{"fill"} shows pointwise 95\% confidence limits for the flexible calibration curve with a gray
area between the lower and upper limits, \code{TRUE} shows pointwise 95\% confidence limits for the flexible calibration curve
 with dashed lines, \code{FALSE} suppresses the confidence limits. Default is \code{"fill"}.}

\item{CL.BT}{\code{TRUE} uses confidence limits based on 2000 bootstrap samples, \code{FALSE} uses closed form confidence limits.
Default is \code{FALSE}.}

\item{nr.knots}{specifies the number of knots for rcs-based calibration curve. The default as well as the highest allowed value is 5.
In case the specified number of knots leads to estimation problems, then the number of knots is automatically reduced to the closest
 value without estimation problems.}

\item{logistic.cal}{\code{T} or \code{TRUE} plots the logistic calibration curve, \code{F} or \code{FALSE} suppresses this curve.
Default is \code{FALSE}.}

\item{xlab}{x-axis label, default is \code{"Predicted Probability"}.}

\item{ylab}{y-axis label, default is \code{"Observed proportion"}.}

\item{xlim, ylim}{numeric vectors of length 2, giving the x and y coordinates ranges (see \code{\link{plot.window}})}

\item{m}{
If grouped proportions are desired, average no. observations per group
}

\item{g}{
If grouped proportions are desired, number of quantile groups
}

\item{cuts}{
If grouped proportions are desired, actual cut points for constructing
intervals, e.g. \code{c(0,.1,.8,.9,1)} or \code{seq(0,1,by=.2)}
}

\item{emax.lim}{
Vector containing lowest and highest predicted probability over which to
compute \code{Emax}.
}

\item{legendloc}{if \code{pl=TRUE}, list with components \code{x,y} or vector \code{c(x,y)} for bottom right corner of legend for
curves and points. Default is \code{c(.50, .27)} scaled to lim. Use \code{locator(1)} to use the mouse, \code{FALSE} to suppress legend.}

\item{statloc}{the "abc" of model performance (Steyerberg et al., 2011)-calibration intercept, calibration slope,
and c statistic-will be added to the plot, using statloc as the upper left corner of a box (default is c(0,.85).
You can specify a list or a vector. Use locator(1) for the mouse, \code{FALSE} to suppress statistics. This is plotted after
the curve legends.}

\item{dostats}{specifies whether and which performance measures are shown in the figure.
\code{TRUE} shows the \code{"abc"} of model performance (Steyerberg et al., 2011): calibration intercept, calibration slope,
 and c statistic. \code{TRUE} is default.
 \code{FALSE} suppresses the presentation of statistics in the figure. A \code{c()} list of specific stats shows the specified
 stats. The key stats which are also mentioned in this paper are \code{"C (ROC)"} for the c statistic, \code{"Intercept"} for the
 calibration intercept, \code{"Slope"} for the calibration slope, and \code{"ECI"} for the estimated calibration index
 (Van Hoorde et al, 2015). The full list of possible statistics is taken from \code{\link{val.prob}}
 and augmented with the estimated calibration index: \code{"Dxy", "C (ROC)", "R2", "D", "D:Chi-sq", "D:p", "U", "U:Chi-sq",
  "U:p", "Q", "Brier", "Intercept", "Slope", "Emax", "Brier scaled", "Eavg", "ECI"}. These statistics are always returned by the function.}

\item{roundstats}{specifies the number of decimals to which the statistics are rounded when shown in the plot. Default is 2.}

\item{riskdist}{
Defaults to \code{"calibrated"} to plot the relative frequency distribution of
calibrated probabilities after dividing into 101 bins from \code{lim[1]} to
\code{lim[2]}.
Set to \code{"predicted"} to use raw assigned risk, \code{FALSE} to omit risk distribution.
Values are scaled so that highest bar is \code{0.15*(lim[2]-lim[1])}.
}

\item{cex, cex.leg}{controls the font size of the statistics (\code{cex}) or plot legend (\code{cex.leg}). Default is 0.75}

\item{connect.group}{
Defaults to \code{FALSE} to only represent group fractions as triangles.
Set to \code{TRUE} to also connect with a solid line.
}

\item{connect.smooth}{Defaults to \code{TRUE} to draw smoothed estimates using a line. Set to \code{FALSE} to instead use dots at individual estimates}

\item{g.group}{
number of quantile groups to use when \code{group} is given and variable is
numeric.
}

\item{evaluate}{
number of points at which to store the \code{lowess}-calibration curve.
Default is 100.  If there are more than \code{evaluate} unique predicted
probabilities, \code{evaluate} equally-spaced quantiles of the unique
predicted probabilities, with linearly interpolated calibrated values,
are retained for plotting (and stored in the object returned by
\code{val.prob}.
}

\item{nmin}{
applies when \code{group} is given.  When \code{nmin} \eqn{> 0}, \code{val.prob} will not
store coordinates of smoothed calibration curves in the outer tails,
where there are fewer than \code{nmin} raw observations represented in
those tails.  If for example \code{nmin}=50, the \code{plot} function will only
plot the estimated calibration curve from \eqn{a} to \eqn{b}, where there are
50 subjects with predicted probabilities \eqn{< a} and \eqn{> b}.
\code{nmin} is ignored when computing accuracy statistics.
}

\item{d0lab, d1lab}{controls the labels for events and non-events (i.e. outcome y) for the histograms.
Defaults are \code{d1lab="1"} for events and \code{d0lab="0"} for non-events.}

\item{cex.d01}{controls the size of the labels for events and non-events. Default is 0.7.}

\item{dist.label}{controls the horizontal position of the labels for events and non-events. Default is 0.04.}

\item{line.bins}{controls the horizontal (y-axis) position of the histograms. Default is -0.05.}

\item{dist.label2}{controls the vertical distance between the labels for events and non-events. Default is 0.03.}

\item{cutoff}{puts an arrow at the specified risk cut-off(s). Default is none.}

\item{las}{controls whether y-axis values are shown horizontally (1) or vertically (0).}

\item{length.seg}{controls the length of the histogram lines. Default is \code{1}.}

\item{y.intersp}{character interspacing for vertical line distances of the legend (\code{\link{legend}})}

\item{col.ideal}{controls the color of the ideal line on the plot. Default is \code{"grey"}.}

\item{lwd.ideal}{controls the line width of the ideal line on the plot. Default is \code{1}.}

\item{...}{arguments to be passed to \code{\link{plot}}, see \code{\link{par}}}
}
\value{
A vector containing performance measures of calibration is returned.
}
\description{
The function val.prob.ci.2 is an adaptation of \code{\link{val.prob}} from Frank Harrell's rms package,
\url{https://cran.r-project.org/web/packages/rms/rms.pdf}. Hence, the description of some of the functions of \code{val.prob.ci.2}
come from the the original \code{\link{val.prob}}.
\cr \cr The key feature of \code{val.prob.ci.2} is the generation of logistic and flexible calibration curves and related statistics.
When using this code, please cite: Van Calster, B., Nieboer, D., Vergouwe, Y., De Cock, B., Pencina, M.J., Steyerberg
E.W. (2016). A calibration hierarchy for risk models was defined: from utopia to empirical data. Journal of Clinical Epidemiology, 74, pp. 167-176
}
\examples{
# simulated data
x1 <- as.matrix(rnorm(500))
x2 <- as.matrix(rnorm(500))
x3 <- as.matrix(rnorm(500))
lp0=0.5*x1+1.2*x2+0.75*x3
p0true=exp(lp0)/(1+exp(lp0))
y <-rbinom(500,1,p0true)
data.0 <- data.frame(y,x1,x2,x3)

# fit logistic model
fit.lrm <- lrm(y~x1+x2+x3,data=data.0)
pred.lrm <- predict(fit.lrm,type="fitted")

# calibration plot
val.prob.ci.2(pred.lrm,y)
}
\references{
Van Calster, B., Nieboer, D., Vergouwe, Y., De Cock, B., Pencina M., Steyerberg E.W. (2016). A calibration hierarchy for risk models was defined: from utopia to empirical data. Journal of Clinical Epidemiology, 74, pp. 167-176

Steyerberg, E.W.Van Calster, B., Pencina, M.J. (2011). Performance measures for prediction models and markers : evaluation of predictions and classifications. Revista Espanola de Cardiologia, 64(9), pp. 788-794

Van Hoorde, K., Van Huffel, S., Timmerman, D., Bourne, T., Van Calster, B. (2015). A spline-based tool to assess and visualize the calibration of multiclass risk predictions. Journal of Biomedical Informatics, 54, pp. 283-93
}

