% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valProbCluster.R
\name{valProbCluster}
\alias{valProbCluster}
\title{Calibration performance with cluster adjustment (ggplot version)}
\usage{
valProbCluster(
  data = NULL,
  p,
  y,
  cluster,
  plot = TRUE,
  approach = c("MIXC", "CGC", "MAC2"),
  cl.level = 0.95,
  xlab = "Predicted probability",
  ylab = "Observed proportion",
  grid_l = 100,
  rangeGrid = range(p),
  ...
)
}
\arguments{
\item{data}{Optional data frame containing the variables \code{p}, \code{y},
and \code{cluster}. If supplied, variable names should be given without
quotation marks.}

\item{p}{Predicted probabilities (numeric vector) or name of the column in
\code{data}.}

\item{y}{Binary outcome (0/1; numeric, integer, or logical) or name of the
column in \code{data}.}

\item{cluster}{Cluster identifier (factor, character, or integer) or name of
the column in \code{data}.}

\item{plot}{Logical. If \code{TRUE}, a plot will be produced by the chosen
subfunction.}

\item{approach}{Character string specifying the calibration method to use.
Must be one of:
\itemize{
  \item \code{"CGC"}: Clustered Generalized Calibration
  \item \code{"MAC2"}: Marginal Calibration (2nd version)
  \item \code{"MIXC"}: Mixed-effects Calibration
}
Defaults to \code{"MIXC"}.}

\item{cl.level}{the confidence level for the calculation of the confidence interval. Default is \code{0.95}.}

\item{xlab}{Label for the x-axis of the plot (default: \code{"Predicted probability"}).}

\item{ylab}{Label for the y-axis of the plot (default: \code{"Observed proportion"}).}

\item{grid_l}{Integer. Number of points in the probability grid for plotting
(default: \code{100}).}

\item{rangeGrid}{the range of the grid. Default is \code{range(p)}.}

\item{...}{Additional arguments passed to the selected subfunction
(\code{\link{CGC}}, \code{\link{MAC2}} and \code{\link{MIXC}}).}
}
\value{
An object of class \code{"valProbCluster"} containing:
\itemize{
  \item \code{call}: The matched call.
  \item \code{approach}: The chosen approach.
  \item \code{cl.level}: the confidence level used.
  \item \code{grid}: Probability grid used for plotting.
  \item \code{ggplot}: A \code{ggplot} object if returned by the subfunction,
        otherwise \code{NULL}.
  \item \code{results}: Results from the chosen subfunction.
  \item \code{labels}: A list with \code{xlab} and \code{ylab}.
}
}
\description{
This function evaluates calibration performance of predicted probabilities
while accounting for clustering. It supports multiple approaches
(`"CGC"`, `"MAC2"`, `"MIXC"`) and returns both results and a `ggplot` object.
Additional arguments can be passed flexibly to the chosen subfunction.
}
\details{
The function internally calls one of the following subfunctions:
\itemize{
  \item \code{CGC(preds, y, cluster, plot, ...)}
  \item \code{MAC2(preds, y, cluster, plot, grid, ...)}
  \item \code{MIXC(preds, y, cluster, plot, CI, grid, ...)}
}

Extra arguments supplied via \code{...} are passed directly to these
subfunctions, providing flexibility in controlling their behavior. For a list
of available arguments for each subfunction, refer to their respective
documentation.
}
\examples{
library(lme4)
data("clustertraindata")
data("clustertestdata")
mFit           = glmer(y ~ x1 + x2 + x3 + x5 + (1 | cluster), data = clustertraindata, family = "binomial")
preds          = predict(mFit, clustertestdata, type = "response", re.form = NA)
y              = clustertestdata$y
cluster        = clustertestdata$cluster
valClusterData = data.frame(y = y, preds = preds, center = cluster)

# Assess calibration performance
Results  = valProbCluster(
p = valClusterData$preds, y = valClusterData$y, cluster = valClusterData$center,
plot = TRUE,
approach = "MIXC", method = "slope", grid_l = 100
)
Results

}
\seealso{
\code{\link{CGC}}, \code{\link{MAC2}} and \code{\link{MIXC}}
}
